pyproject.toml
[project]
name = "remarkable-ai"
version = "0.1.0"
description = ""
authors = [{ name = "Joep Buhre", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "uvicorn (>=0.34.0,<0.35.0)", "imap-tools (>=1.10.0,<2.0.0)", "python-dotenv (>=1.1.0,<2.0.0)", "loguru (>=0.7.3,<0.8.0)", "openai (>=1.69.0,<2.0.0)", "pypandoc (>=1.15,<2.0)", "python-multipart (>=0.0.20,<0.0.21)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "remarkable-ai"
packages = [{ include = "remarkable_ai", from = "src" }]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

